budget-service/requirements.txt
Django==5.2.2
mysqlclient

djangorestframework
corsheaders
